OpInfo: Add outputs_uninitialized property
Currently all empty-like functions manually skip lots of OpInfo tests
because their tensor ouputs are uninitialized and so can't be compared
with `assertEqual`. Instead, this adds an `assertOutputsMatch` method
which will check the `outputs_uninitialized` property and if it's set
it will only compare tensor meta-data as if the tensor was a
meta-device tensor.

To support the metadata-only comparison, I've split the
`TensorLikePair` class into `TensorMetaPair` for comparing attributes
and `TensorLikePair` which is derived from `TensorMetaPair` and adds
the value-comparisons.

ghstack-source-id: b310ad90bce1de128855f6136ce4a33c715e7021
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84087