Improve repeat_interleave with scalar repeat value
`repeat_interleave_symint` is currently implemented by guarding on the `SymInt`
and converting it to a tensor to pass to the Tensor overload. This instead
implements it as a copy of an expanded tensor, which can be done without guards
and is also much more efficient in eager mode to boot.

ghstack-source-id: 9b9027870d635f1e47c0e6ef1ec129456fb58629
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102570