[inductor] Handle aten.full's dtype in the decomposition
In the lowering we don't have `SymFloat` and `SymInt`, we just have `sympy.Expr`
so it is impossible to accurately determine the expected dtype of a `full` call.
For example, `sym_float(int_expr)` has `is_integer=True` but should be treated
as a float. In the decomposition though, we can get this right.

ghstack-source-id: a5b8e323cb166621f796109d6bad4cfcc903a8de
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108443