[minifier] Add config flag to ignore non-fp values
When minifying, the after-aot minifier ignores non-floating values by
default but does check them when running the the initial graph dump step.
This means we may capture a graph that doesn't fail the tester and doesn't have
any meaningful divergence.

For example, the derivative of `elu(x)` depends on `x > 0` so this value is
saved for backwards and so becomes a graph output. However, the difference
between `FLT_MIN` and `0` in `x` is now enough to trigger an accuracy failure.

I fix this by adding a config variable and environment variable to ignore these
non floating point values.

ghstack-source-id: f6eecaad36c7a6637a6a646aa9e49baaa224ac45
Pull Request resolved: https://github.com/pytorch/pytorch/pull/123006