[inductor] Take absolute value of strides when picking loop order
The stride hint is found by comparing the value of the indexing expression
evaluated at `idx` set to all zeros and at `idx[dim] = 1`. This causes a problem
for padded inputs where 0 and 1 are still in the padded region.

In particular, for reflection padding this causes the stride to be negative.

ghstack-source-id: 18109437e7f6e23a215948d7b20cccb5f70e9c41
Pull Request resolved: https://github.com/pytorch/pytorch/pull/127425